Solving #2

+++ Makefile
@@ -13,9 +13,10 @@ clean:
 	for i in $(SUB); do $(MAKE) -C $$i clean; done
 
 install:
-	sudo dnf -y install R texlive-framed
-	R -e 'install.packages("knitr", dependencies = TRUE)'
-	R -e 'install.packages("ggplot2")'
+	if [ -f /etc/redhat-release ]; then	sudo dnf -y install R texlive-framed; elif [ -f /etc/debian_version ]; then sudo apt-get install r-base r-base-dev texlive r-cran-rgl r-cran-ggplot2 -y; fi
+	R -e 'install.packages("knitr", dependencies = TRUE, repos="http://cran.rstudio.com/")'
+	R -e 'install.packages("ggplot2", dependencies = TRUE, repos="http://cran.rstudio.com/")'
+	wget https://raw.githubusercontent.com/geneura/biblio/master/geneura.bib
 
 %.tex: %.Rnw
 	R -e 'library(knitr);knit("$<")'
